Add files via upload
These are some of the books we were able to add to our digital shelves, however, we are still trying to make applicable changes to ensure one can have sources for all books from all the authors. we will soon be able to link an API to enable broader searches, as we try to debug the problem.